requirements.txt
absl-py
alog
autoreject
beautifulsoup4
bleak
clipboard
certstore
coverage
dateparser
dominate
h5io
imageio
joblib
lspopt
lz4
mat73
mne
mne-bids-pipeline
monitorcontrol
natsort
networkx
numba
numpy
numpyencoder
opencv-python
pandas
pingouin
pip-system-certs
prettytable
pybids
pybind11
pyedflib
pyexcel
pyexcel-ods
pyexcel-ods3
pygame
pyglet
pymupdf
pytablewriter
python-picard
python-pptx
requests
scikit-learn
scipy
seaborn
sleep_utils
standard-imghdr
statsmodels
spyder-kernels
telegram-send
truststore
tqdm
wmi
